pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "financial_ingestion"
version = "0.0.1"
description = "Financial data ingestion pipeline"
authors = [{name = "Neha Prasad", email = "18.prasadneha@gmail.com"}]
dependencies = [
    "yfinance==0.2.52",
    "pydantic==2.10.6",
    "pymongo==4.10.1",
    "pyyaml==6.0.2",
    "python-dotenv==1.0.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]